Refactor can_vectorize & simd_abi::native
Remove the duplication of logic that determines valid vector widths. The
new function __vectorized_sizeof<_Tp> returns the native SIMD register
sizeof for the given _Tp and the ISA extensions active in the current
TU. Use this function to implement simd_abi::native and all checks for
validity of ABI tags.

	* include/experimental/bits/simd.h: New function
	__vectorized_sizeof<_Tp>. Use __vectorized_sizeof to implement
	__determine_native_abi<_Tp> (simd_abi::native<_Tp>).
	* include/experimental/bits/simd_builtin.h: Remove
	__can_vectorize<_Tp> and replace all occurrences with checks
	calling __vectorized_sizeof<_Tp>.