Handle UB cases in left and right bitshift
The behavior is undefined if the right operand is negative, or greater
than or equal to the width of the promoted left operand. In the vector
case there is no promoted left operand so the compiler may assume UB
where there wouldn't be one with scalar types. Avoid that UB but
reinstate the UB on shifting out of the promoted type.

	* include/experimental/bits/simd_builtin.h: Handle UB cases in
	left and right bitshift.